Change the fSecs field of plUnifiedTime from UInt32 to time_t.
This fixes various date formatting problems when building on a system where time_t is 64-bit (e.g. Visual Studio 2010), and, as a bonus, extends the range past 2038 on such systems.

The wire protocol is left at 32-bit for now, we might change that to 64 when other reasons to break compatibility have accumulated.

+++ Sources/Plasma/FeatureLib/pfPython/pyDniInfoSource.cpp
@@ -55,10 +55,10 @@ UInt32 pyDniInfoSource::GetAgeTime( void ) const
     if (!node)
         return 0;
     
-    unsigned result;
+    UInt32 result;
     VaultAgeInfoNode ageInfo(node);
     if (const plUnifiedTime * utime = ageInfo.GetAgeTime())
-        result = utime->GetSecs();
+        result = (UInt32)utime->GetSecs();
     else
         result = 0;
     node->DecRef();
